refactor(plugin-sdk): replace duplicate approval exports
Expose terminal presentation only through approval-runtime and preserve both
precedence orders. Retire three unused broad bindings, make the session
predicate private, and retain the narrow account helpers used by plugins.

Proof: canonical SDK export/surface checks (4446 exports, 2630 callables),
48 files/547 focused tests, 3 filter tests, and targeted typechecks passed.
