+++ docs/plugins/sdk-channel-plugins.md
@@ -648,13 +648,18 @@ Other approval helpers:
   their own validation.
 - Use `formatChannelApprovalResolvedLabel` and
   `buildSystemAgentApprovalResolvedText` from
-  `openclaw/plugin-sdk/approval-handler-runtime` for terminal presentation.
+  `openclaw/plugin-sdk/approval-runtime` for terminal presentation.
   Rich labels preserve application-status precedence; prose preserves denial
   precedence, because a denied system change can also report `not-applied`.
   Both prioritize cancellation. Pass a decision formatter for transport-specific
   label spelling, and prepare any bounded operation summary before building prose.
   Use `formatApprovalDecisionLabel` for a recorded decision without implying
   application completion.
+- Approval account lookup helpers `resolveApprovalRequestAccountId` and
+  `resolveApprovalRequestChannelAccountId` use `approval-native-runtime`. Their
+  duplicate `approval-runtime` exports and its unused
+  `matchesApprovalRequestSessionFilter` export have been retired. The core
+  implementations are unchanged.
 - Use `createNativeApprovalChannelRouteGates` from
   `openclaw/plugin-sdk/approval-native-runtime` when a channel supports both
   session-origin native delivery and explicit approval forwarding targets. The
