+++ src/plugin-sdk/approval-runtime.ts
@@ -31,11 +31,7 @@ export {
   resolveExecApprovalSessionTarget,
   type ExecApprovalSessionTarget,
 } from "../infra/exec-approval-session-target.js";
-export {
-  doesApprovalRequestMatchChannelAccount,
-  resolveApprovalRequestAccountId,
-  resolveApprovalRequestChannelAccountId,
-} from "../infra/approval-request-account-binding.js";
+export { doesApprovalRequestMatchChannelAccount } from "../infra/approval-request-account-binding.js";
 export {
   buildPluginApprovalExpiredMessage,
   buildPluginApprovalRequestMessage,
@@ -67,7 +63,6 @@ export {
 export { resolveApprovalApprovers } from "./approval-approvers.js";
 export {
   matchesApprovalRequestFilters,
-  matchesApprovalRequestSessionFilter,
   type ApprovalRequestFilterInput,
 } from "../infra/approval-request-filters.js";
 export {
@@ -76,3 +71,8 @@ export {
   buildPluginApprovalPendingReplyPayload,
   buildPluginApprovalResolvedReplyPayload,
 } from "./approval-renderers.js";
+export {
+  buildSystemAgentApprovalResolvedText,
+  formatApprovalDecisionLabel,
+  formatChannelApprovalResolvedLabel,
+} from "./approval-terminal.js";
